+++ Changelog.md
@@ -4,6 +4,12 @@
 
 **Deprecation Warning:** *Currently the registry returns `OrderedDict` objects when you're querying for regions or subregions. Expect that the next release will preferrably return plain'ol' `dict` objects. If your scripts rely on the order of the objects returned, you'll have to sort them yourself.*
 
+### New Calendar
+
+- Added Turkey by @tayyipgoren (#371).
+
+### Other changes
+
 - Change registry mechanism to avoid circular imports (#288).
 - Internal: Added a "Release" section to the Pull Request template.
 - Internal: Added advices on the Changelog entry in the Contributing document.